agent-service/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx~=0.25.2
fastapi==0.104.1
uvicorn==0.24.0
claude-agent-sdk==0.1.5
supabase==2.0.3
supabase==2.28.0
python-dotenv==1.0.0
httpx>=0.26,<0.29
pydantic==2.12.5
